functions-python/requirements.txt
firebase-functions>=0.1.0
praw>=7.7.0
flask>=2.0.0
python-dotenv>=1.0.0
